requirements.txt
aiosqlite==0.20.0
bcrypt==4.1.3
cryptography==42.0.8
fastapi==0.111.0
gunicorn==23.0.0
httpx==0.27.0
itsdangerous==2.2.0
Jinja2==3.1.4
numpy==1.26.4
passlib[bcrypt]==1.7.4
psutil==5.9.8
pydantic==2.7.4
pydantic-settings==2.3.4
python-json-logger==2.0.7
python-multipart==0.0.18
redis==5.0.7
scikit-learn==1.5.0
SQLAlchemy==2.0.31
uvicorn==0.30.1
aiofiles==23.2.1
aiohttp==3.9.5
pytest-asyncio==0.23.7
